+++ packages/terra-search-field/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 3.77.0 - (January 9, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 3.76.0 - (December 7, 2022)
 
 * Changed